restricted user from access unauthorized pages

+++ base/views.py
@@ -1,6 +1,8 @@
 from django.shortcuts import redirect, render
+from django.http import HttpResponse
 from django.db.models import Q
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate, login, logout
 
@@ -9,7 +11,11 @@ from .forms import RoomForm
 
 # Create your views here.
 
+# Login user
 def loginPage(request):
+
+    if request.user.is_authenticated:
+        return redirect('home')
 
     if request.method == "POST":
         username = request.POST.get('username')
@@ -32,9 +38,12 @@ def loginPage(request):
     context = {}
     return render(request, 'base/login_register.html', context)
 
+
+# Logout user
 def logoutUser(request):
     logout(request)
     return redirect('home')
+
 
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ''
@@ -60,6 +69,7 @@ def room(request, pk):
 
 
 # Room CRUD
+@login_required(login_url='login')
 def create_room(request):
     form = RoomForm()
     if request.method == "POST":
@@ -74,9 +84,14 @@ def create_room(request):
     return render(request, 'base/room_form.html', context)
 
 
+@login_required(login_url='login')
 def update_room(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    # If not the host of the room, then cannot update the room
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here!')
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -90,8 +105,15 @@ def update_room(request, pk):
     }
     return render(request, 'base/room_form.html', context)
 
+
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    # If not the host of the room, then cannot delete the room
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here!')
+
     if request.method == 'POST':
         room.delete()
         return redirect('home')
